# packages/backend/src/controllers/Categories/Categories.DTO.ts
/**
 * Стандартная DTO для категории
 * @tsoaModel
 */
export interface CategoriesDTO {
  /**
   * @isInt
   * @example 1
   */
  id: number;

  /**
   * @example "Процессоры"
   */
  name: string;
}